OWLAPIAdaptor disbanded, logic moved to SyncReasoner
